setting the edition size to 1000

+++ web/src/lib/graphql.ts
@@ -245,7 +245,7 @@ export type CreateBuyNowItemType = {
 };
 
 export type CreateMultipleTracksInput = {
-  amount: Scalars['Float'];
+  amount: Scalars['Int'];
   track: CreateTrackInput;
 };
